requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
greenlet==1.1.2
idna==3.3
python-telegram-bot==13.12
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.37
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
